+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - atoms-dropdown: added single-select dropdown component.
+- molecules-search-search_field: added search field form component.
+- atoms-checkbox: added checkbox form component.
 
 ### Removed